ExchangeProcessを移管
